+++ SolrNet.Tests/MultivaluedMappedToCollectionRuleTests.cs
@@ -40,7 +40,7 @@ namespace SolrNet.Tests {
             var schema = solrSchemaParser.Parse(schemaXmlDocument);
 
             var validationResults = schemaManager.EnumerateValidationResults(typeof (SchemaMappingTestDocument), schema).ToList();
-            Assert.Equal(1, validationResults.Count);
+            Assert.Single(validationResults);
             Console.WriteLine(validationResults[0].Message);
         }
 
@@ -56,7 +56,7 @@ namespace SolrNet.Tests {
             var schema = solrSchemaParser.Parse(schemaXmlDocument);
 
             var validationResults = schemaManager.EnumerateValidationResults(typeof (SchemaMappingTestDocument), schema).ToList();
-            Assert.Equal(0, validationResults.Count);
+            Assert.Empty(validationResults);
         }
 
         [Fact]
@@ -71,7 +71,7 @@ namespace SolrNet.Tests {
             var schema = solrSchemaParser.Parse(schemaXmlDocument);
 
             var validationResults = schemaManager.EnumerateValidationResults(typeof (SchemaMappingTestDocument), schema).ToList();
-            Assert.Equal(1, validationResults.Count);
+            Assert.Single(validationResults);
             Console.WriteLine(validationResults[0].Message);
         }
     }